[Refactor] Rename FlyGpuStreamInject -> FlyGpuToLLVM to match pass name
Align directory, file, and CMake library names with the actual pass name
fly-gpu-to-llvm. Also register the pass and dialect extensions in fly-opt
to enable FileCheck testing.
