+++ lib/CAPI/Dialect/FlyROCDL/CMakeLists.txt
@@ -3,5 +3,5 @@ add_mlir_public_c_api_library(MLIRCPIFlyROCDL
   LINK_LIBS PUBLIC
   MLIRFlyROCDLDialect
   MLIRFlyToROCDL
-  MLIRFlyGpuStreamInject
+  MLIRFlyGpuToLLVM
 )
